fix(recommendations): order half-numbered series entries by fractional position
A 200-book run recommended Expanse novellas "Drive (#2.7)", "Gods of Risk
(#2.5)", and "The Vital Abyss (#5.5)" out of series order, ahead of the
unread next book "Caliban's War (#2)".

Series-position parsing was integer-only: the title regex captured `\d+`
and stopped at the decimal point, and the metadata path's `int()` raised on
"2.7". Fractional novellas therefore parsed as non-series items, dodged the
"too far ahead" suppression, and floated up on raw similarity.

Capture an optional fractional part in the series patterns, parse positions
with `float()`, and migrate series tracking from `set[int]` to `set[float]`
throughout. Rework `should_recommend_item`'s gap logic from integer
`range()` scans to a set-based "earliest unconsumed position" rule so a
`#2.5` novella waits until its preceding integer book is read. A
`math.isfinite` guard rejects malformed "inf"/"nan" metadata positions that
`float()` would otherwise accept.

+++ src/recommendations/engine.py
@@ -468,7 +468,7 @@ class RecommendationEngine:
     def _apply_series_filtering(
         self,
         pipeline_scored: list[ScoredCandidate],
-        series_tracking: dict[str, set[int]],
+        series_tracking: dict[str, set[float]],
         unconsumed_items: list[ContentItem],
     ) -> list[ScoredCandidate]:
         """Filter and substitute candidates based on series ordering rules.
@@ -707,7 +707,7 @@ class RecommendationEngine:
         all_consumed_items: list[ContentItem],
         unconsumed_items: list[ContentItem],
         count: int,
-        series_tracking: dict[str, set[int]],
+        series_tracking: dict[str, set[float]],
     ) -> None:
         """Enhance recommendations with LLM-generated reasoning.
 
@@ -840,7 +840,7 @@ class RecommendationEngine:
     def _build_fallback_recommendations(
         self,
         unconsumed_items: list[ContentItem],
-        series_tracking: dict[str, set[int]],
+        series_tracking: dict[str, set[float]],
         count: int,
     ) -> list[dict[str, Any]]:
         """Build fallback recommendations when no scored recommendations exist.
